Feature logging api

+++ src/Services/Logging.API/Services/ScriptInstanceService.cs
@@ -9,13 +9,19 @@ using Dapper.Contrib.Extensions;
 using EESLP.Services.Logging.API.Entities;
 using EESLP.Services.Logging.API.ViewModel;
 using Dapper;
+using EESLP.BuildingBlocks.Resilence.Http;
+using EESLP.Services.Logging.API.Infrastructure.Exceptions;
 
 namespace EESLP.Services.Logging.API.Services
 {
     public class ScriptInstanceService : BaseService, IScriptInstanceService
     {
-        public ScriptInstanceService(IOptions<DatabaseOptions> options) : base(options)
+        private readonly IHttpApiClient _http;
+        private readonly ApiOptions _apiOptions;
+        public ScriptInstanceService(IOptions<DatabaseOptions> databaseOptions, IHttpApiClient http, IOptions<ApiOptions> apiOptions) : base(databaseOptions)
         {
+            _http = http;
+            _apiOptions = apiOptions.Value;
         }
 
         public IEnumerable<ScriptInstance> GetAllScriptInstances()
@@ -49,6 +55,10 @@ namespace EESLP.Services.Logging.API.Services
         {
             using (var db = Connection)
             {
+                if (!checkIfScriptExists(scriptInstance.ScriptId) || !checkIfHostExists(scriptInstance.HostId))
+                {
+                    throw new EntityNotFoundException();
+                }
                 db.Open();
                 UpdateAuditableFields(scriptInstance);
                 return db.Update<ScriptInstance>(scriptInstance);
@@ -68,6 +78,9 @@ namespace EESLP.Services.Logging.API.Services
         {
             using (var db = Connection)
             {
+                if (!checkIfScriptExists(scriptInstance.ScriptId) || !checkIfHostExists(scriptInstance.HostId)) {
+                    throw new EntityNotFoundException();
+                }
                 db.Open();
                 UpdateAuditableFields(scriptInstance, true);
                 return (int)db.Insert<ScriptInstance>(scriptInstance);
@@ -78,6 +91,10 @@ namespace EESLP.Services.Logging.API.Services
         {
             using (var db = Connection)
             {
+                if (!checkIfScriptExists(scriptid))
+                {
+                    throw new EntityNotFoundException();
+                }
                 db.Open();
                 db.Execute("DELETE FROM EESLP.ScriptInstance WHERE ScriptId = @Id ", new
                 {
@@ -90,12 +107,28 @@ namespace EESLP.Services.Logging.API.Services
         {
             using (var db = Connection)
             {
+                if (!checkIfHostExists(hostid))
+                {
+                    throw new EntityNotFoundException();
+                }
                 db.Open();
                 db.Execute("DELETE FROM EESLP.ScriptInstance WHERE HostId = @Id", new
                 {
                     Id = hostid
                 });
             }
+        }
+
+        // Help functions
+        private bool checkIfScriptExists(int scriptid)
+        {
+            var script = _http.GetStringAsync(_apiOptions.ScriptsApiUrl + "/api/Scripts/" + scriptid).Result;
+            return script != null && script != "";
+        }
+        private bool checkIfHostExists(int hostid)
+        {
+            var script = _http.GetStringAsync(_apiOptions.ScriptsApiUrl + "/api/Hosts/" + hostid).Result;
+            return script != null && script != "";
         }
     }
 }
